refactor: ui.common 配下のファイルを :core:ui モジュールに移行した。
