+++ core/ui/src/main/res/drawable/jetbrains.png
